backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
chromadb
pandas
google-generativeai
openai
python-dotenv
requests
sentence-transformers
nltk
